+++ source/rafcon/gui/views/execution_history.py
@@ -43,10 +43,13 @@ class ExecutionHistoryView(View, gtk.ScrolledWindow):
         reload_button.set_border_width(constants.BUTTON_BORDER_WIDTH)
         clean_button = gtk.Button("Clean history")
         clean_button.set_border_width(constants.BUTTON_BORDER_WIDTH)
+        open_separately_button = gtk.Button("Open separately")
+        open_separately_button.set_border_width(constants.BUTTON_BORDER_WIDTH)
 
         button_box = gtk.HBox()
         button_box.pack_end(reload_button, False, True, 0)
         button_box.pack_end(clean_button, False, True, 0)
+        button_box.pack_end(open_separately_button, False, True, 0)
 
         history_vbox = gtk.VBox()
         history_vbox.pack_end(button_box, False, True, 0)
@@ -61,4 +64,5 @@ class ExecutionHistoryView(View, gtk.ScrolledWindow):
         self['history_tree'] = history_tree
         self['reload_button'] = reload_button
         self['clean_button'] = clean_button
+        self['open_separately_button'] = open_separately_button
         self.top = 'history_view'